Added travis integration for sdk tests.
Through arguments while executing maven user and password values can now
be specified to the program. The CI can add it's environment variables to
it. If the CI has a value for key user and a value for key pwd it can run
the tests that need to connect with the Tygron server.